requirements.txt
altair
annotated-types
appnope
asttokens
attrs
backcall
beautifulsoup4
bleach
blinker
cachetools
certifi
charset-normalizer
click
decorator
defusedxml
docopt
executing
fastjsonschema
gitdb
GitPython
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-generativeai
googleapis-common-protos
grpcio
grpcio-status
httplib2
idna
importlib_metadata
ipython
jedi
Jinja2
jsonschema
jsonschema-specifications2023.12.1
jupyter_client 
jupyter_core 
jupyterlab_pygments 
markdown-it-py 
MarkupSafe
matplotlib-inline
mdurl
mistune
nbclient
nbconvert
nbformat
numpy
packaging
pandas
pandocfilters
parso
pdf2image
pexpect
pickleshare
pillow
pipreqs
platformdirs
prompt_toolkit
proto-plus
protobuf
ptyprocess
pure_eval
pyarrow
pyasn1
pyasn1_modules
pydantic
pydantic_core
pydeck
Pygments
pyparsing
python-dateutil
python-dotenv
pytz
pyzmq
referencing
requests
rich
rpds-py
rsa
six
smmap
soupsieve
stack-data
streamlit
tenacity
tinycss2
toml
toolz
tornado
tqdm
traitlets
typing_extensions
tzdata
uritemplate
urllib3
wcwidth
webencodings
yarg
zipp
